fastApiProject/pyproject.toml
[tool.poetry]
name = "\\"
version = "0.1.0"
description = ""
authors = ["yo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
alembic = "^1.13.1"
websockets = "^12.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.25"
install = "^1.3.5"
pydantic-settings = "^2.1.0"
aiohttp = "^3.9.3"
aioredis = "^2.0.1"
nats-python = "^0.8.0"
nats-py = "^2.7.0"
uvicorn = "^0.27.0.post1"
aiomultiprocess = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
